ws-rasmoo-plus/src/test/java/com/client/ws/rasmooplus/presentation/controllers/UserControllerTest.java
package com.client.ws.rasmooplus.presentation.controllers;

import com.client.ws.rasmooplus.domain.entities.jpa.UserEntity;
import com.client.ws.rasmooplus.useCases.UserUseCase;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.autoconfigure.orm.jpa.AutoConfigureDataJpa;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.HttpMethod;
import org.springframework.http.MediaType;
import org.springframework.mock.web.MockMultipartFile;
import org.springframework.test.context.ActiveProfiles;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMultipartHttpServletRequestBuilder;

import java.io.FileInputStream;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.multipart;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@AutoConfigureDataJpa
@AutoConfigureTestDatabase
@WebMvcTest(UserController.class)
@AutoConfigureMockMvc(addFilters = false)
@ActiveProfiles(profiles = "test")
class UserControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private UserUseCase userUseCase;

    @Test
    void given_uploadPhoto_when_receiveMulPartFile_then_return200() throws Exception {
        FileInputStream fis = new FileInputStream("src/test/resources/static/test_image.webp");
        MockMultipartFile file = new MockMultipartFile("file", "test_image.webp", MediaType.MULTIPART_FORM_DATA_VALUE, fis);

        Mockito.when(userUseCase.uploadPhoto(1L, file)).thenReturn(new UserEntity());

        MockMultipartHttpServletRequestBuilder builder = multipart("/user/1/upload-photo");
        builder.with(request -> {
            request.setMethod(HttpMethod.PATCH.name());
            return request;
        });

        mockMvc.perform(builder.file(file))
                .andExpect(status().isOk());
    }

    @Test
    void given_downloadPhoto_when_thereIsPhotoInDatabase_then_return200() throws Exception {

        mockMvc.perform(get("/user/1/photo"))
                .andExpect(status().isOk());
        Mockito.verify(userUseCase, Mockito.times(1)).downloadPhoto(1L);
    }
}